Replace *Port connection functions with Settings equivalents
This fixes Issue 1 by allowing the user to set the maximum line length.
It does slightly break the similarity with HaskellNet, which uses the
*Port suffix to specify a port for each connection, but it seems like
the only sensible way to do it, really.  I opted not to keep the *Port
functions in because the proliferation of functions doing basically the
same thing would be too much, I think.